requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.2
black==22.6.0
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dnslib==0.9.19
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.75.2
fastapi-utils==0.2.1
filelock==3.4.0
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
httpie==3.1.0
idna==3.3
jwcrypto==1.3.1
motor==2.5.1
multidict==6.0.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
Pillow==9.2.0
pipenv==2021.11.23
platformdirs==2.4.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
PyJWT==2.4.0
pymongo==3.12.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jwt==3.3.2
python-multipart==0.0.5
PythonDNS==0.1
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.10.0
virtualenv-clone==0.5.7
websocket==0.2.1
websockets==10.3
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
